Updates to use Bootstrapper to customize ConfigData.
Removes AbstractConfigData* classes.

Creates ConfigServerBootstrapper.java to allow simplified bootstrapping for users.
