+++ spring-cloud-config-client/src/main/java/org/springframework/cloud/config/client/ConfigServiceBootstrapConfiguration.java
@@ -46,6 +46,7 @@ public class ConfigServiceBootstrapConfiguration {
 	private ConfigurableEnvironment environment;
 
 	@Bean
+	@ConditionalOnMissingBean
 	public ConfigClientProperties configClientProperties() {
 		ConfigClientProperties client = new ConfigClientProperties(this.environment);
 		return client;
