--- spring-cloud-config-client/src/test/java/org/springframework/cloud/config/client/TestConfigServerConfigDataLocation.java
@@ -1,29 +1,0 @@
-/*
- * Copyright 2013-2020 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframework.cloud.config.client;
-
-import org.springframework.boot.context.config.Profiles;
-import org.springframework.web.client.RestTemplate;
-
-public class TestConfigServerConfigDataLocation extends AbstractConfigDataLocation {
-
-	public TestConfigServerConfigDataLocation(RestTemplate restTemplate, ConfigClientProperties properties,
-			boolean optional, Profiles profiles) {
-		super(restTemplate, properties, optional, profiles);
-	}
-
-}
